+++ src/misc/changelog.ts
@@ -15,6 +15,7 @@ export const changelog = [
 		"• Added a level 2 to backpack research for Rock Golems",
 		"Features",
 		"• Wolf gold command now accepts an argument for how much gold",
+		"• All bounties are now proximity based",
 		"Refactors",
 		"• Wrapped all TriggerAddAction calls with logging",
 	].join( "\n" ) },